remove duplicate Explore page, use flag instead
