+++ src/pages/App.tsx
@@ -36,7 +36,6 @@ const Collection = lazy(() => import('nft/pages/collection'))
 const Profile = lazy(() => import('nft/pages/profile'))
 const Asset = lazy(() => import('nft/pages/asset/Asset'))
 const AddLiquidity = lazy(() => import('pages/AddLiquidity'))
-const Explore = lazy(() => import('pages/Explore'))
 const RedirectDuplicateTokenIds = lazy(() => import('pages/AddLiquidity/redirects'))
 const RedirectDuplicateTokenIdsV2 = lazy(() => import('pages/AddLiquidityV2/redirects'))
 const MigrateV2 = lazy(() => import('pages/MigrateV2'))
@@ -234,8 +233,9 @@ export default function App() {
                       browserRouterEnabled && hash ? <Navigate to={hash.replace('#', '')} replace /> : <Landing />
                     }
                   />
+                  {/* todo: rename Tokens component to Explore */}
                   {infoExplorePageEnabled ? (
-                    <Route path="explore" element={<Explore />}>
+                    <Route path="explore" element={<Tokens />}>
                       <Route path=":chainName" />
                     </Route>
                   ) : (
