+++ src/pages/Tokens/index.tsx
@@ -1,6 +1,8 @@
 import { Trans } from '@lingui/macro'
-import { InterfacePageName } from '@uniswap/analytics-events'
+import { BrowserEvent, InterfacePageName, SharedEventName } from '@uniswap/analytics-events'
+import { TraceEvent } from 'analytics'
 import { Trace } from 'analytics'
+import { AutoRow } from 'components/Row'
 import { MAX_WIDTH_MEDIA_BREAKPOINT, MEDIUM_MEDIA_BREAKPOINT } from 'components/Tokens/constants'
 import { filterStringAtom } from 'components/Tokens/state'
 import NetworkFilter from 'components/Tokens/TokenTable/NetworkFilter'
@@ -8,8 +10,9 @@ import SearchBar from 'components/Tokens/TokenTable/SearchBar'
 import TimeSelector from 'components/Tokens/TokenTable/TimeSelector'
 import TokenTable from 'components/Tokens/TokenTable/TokenTable'
 import { MouseoverTooltip } from 'components/Tooltip'
+import { useInfoExplorePageEnabled } from 'featureFlags/flags/infoExplore'
 import { useResetAtom } from 'jotai/utils'
-import { useEffect } from 'react'
+import { useEffect, useState } from 'react'
 import { useLocation } from 'react-router-dom'
 import styled from 'styled-components'
 import { ThemedText } from 'theme/components'
@@ -34,25 +37,46 @@ const TitleContainer = styled.div`
   margin-right: auto;
   display: flex;
 `
-const FiltersContainer = styled.div`
+const Nav = styled(AutoRow)`
+  gap: 20px;
+`
+
+const NavItem = styled(ThemedText.MediumHeader)<{ active?: boolean }>`
+  align-items: center;
+  color: ${({ theme, active }) => (active ? theme.neutral1 : theme.neutral2)};
+  cursor: pointer;
+  display: flex;
+  justify-content: space-between;
+  transition: ${({ theme }) => `${theme.transition.duration.medium} ${theme.transition.timing.ease} color`};
+  &:hover {
+    ${({ theme, active }) => !active && `color: ${theme.neutral2}`};
+  }
+`
+const FiltersContainer = styled.div<{ isExplore: boolean }>`
   display: flex;
   gap: 8px;
   height: 40px;
 
   @media only screen and (max-width: ${MEDIUM_MEDIA_BREAKPOINT}) {
-    order: 2;
+    ${({ isExplore }) => !isExplore && 'order: 2;'}
+    ${({ isExplore }) => isExplore && 'justify-content: space-between;'}
   }
 `
-const SearchContainer = styled(FiltersContainer)`
-  margin-left: 8px;
+const DropdownFilterContainer = styled(FiltersContainer)`
+  @media only screen and (max-width: ${MEDIUM_MEDIA_BREAKPOINT}) {
+    justify-content: flex-start;
+  }
+`
+const SearchContainer = styled(FiltersContainer)<{ isExplore: boolean }>`
+  ${({ isExplore }) => !isExplore && 'margin: 8px;'}
   width: 100%;
 
   @media only screen and (max-width: ${MEDIUM_MEDIA_BREAKPOINT}) {
-    margin: 0px;
-    order: 1;
+    ${({ isExplore }) => !isExplore && 'order: 1; margin: 0px;'}
+    ${({ isExplore }) => isExplore && 'justify-content: flex-end;'}
   }
 `
-const FiltersWrapper = styled.div`
+const NavWrapper = styled.div`
   display: flex;
   max-width: ${MAX_WIDTH_MEDIA_BREAKPOINT};
   margin: 0 auto;
@@ -66,37 +90,111 @@ const FiltersWrapper = styled.div`
   }
 `
 
+// todo: copied from MiniPortfolio/index.tsx,, DRY?
+interface Page {
+  title: React.ReactNode
+  key: string
+  component: () => JSX.Element
+  loggingElementName: string
+}
+
+const Pages: Array<Page> = [
+  {
+    title: <Trans>Tokens</Trans>,
+    key: 'tokens', // todo: put this into an enum??
+    component: TokenTable,
+    loggingElementName: 'explore-tokens-tab', // todo: add to InterfaceSectionName @uniswap/analytics-events
+  },
+  {
+    title: <Trans>Pools</Trans>,
+    key: 'pools',
+    component: TokenTable,
+    loggingElementName: 'explore-pools-tab',
+  },
+  {
+    title: <Trans>Transactions</Trans>,
+    key: 'transactions',
+    component: TokenTable,
+    loggingElementName: 'explore-transactions-tab',
+  },
+]
+
 const Tokens = () => {
   const resetFilterString = useResetAtom(filterStringAtom)
   const location = useLocation()
+  const [currentTab, setCurrentTab] = useState(0)
+  const isExplore = useInfoExplorePageEnabled()
 
   useEffect(() => {
     resetFilterString()
   }, [location, resetFilterString])
 
+  const { component: Page, key: currentKey } = Pages[currentTab]
+
   return (
-    <Trace page={InterfacePageName.TOKENS_PAGE} shouldLogImpression>
+    // TODO: add 'explore-page' to InterfacePageName in @uniswap/analytics-events
+    <Trace page={isExplore ? 'explore-page' : InterfacePageName.TOKENS_PAGE} shouldLogImpression>
       <ExploreContainer>
-        <TitleContainer>
-          <MouseoverTooltip
-            text={<Trans>This table contains the top tokens by Uniswap volume, sorted based on your input.</Trans>}
-            placement="bottom"
-          >
-            <ThemedText.LargeHeader>
-              <Trans>Top tokens on Uniswap</Trans>
-            </ThemedText.LargeHeader>
-          </MouseoverTooltip>
-        </TitleContainer>
-        <FiltersWrapper>
-          <FiltersContainer>
-            <NetworkFilter />
-            <TimeSelector />
-          </FiltersContainer>
-          <SearchContainer>
-            <SearchBar />
-          </SearchContainer>
-        </FiltersWrapper>
-        <TokenTable />
+        {/* TODO: add graphs to explore page */}
+        {isExplore ? (
+          <div></div>
+        ) : (
+          <TitleContainer>
+            <MouseoverTooltip
+              text={<Trans>This table contains the top tokens by Uniswap volume, sorted based on your input.</Trans>}
+              placement="bottom"
+            >
+              <ThemedText.LargeHeader>
+                <Trans>Top tokens on Uniswap</Trans>
+              </ThemedText.LargeHeader>
+            </MouseoverTooltip>
+          </TitleContainer>
+        )}
+        <NavWrapper>
+          {isExplore && (
+            <Nav data-testid="explore-navbar">
+              {Pages.map(({ title, loggingElementName, key }, index) => {
+                const handleNavItemClick = () => {
+                  setCurrentTab(index)
+                }
+                return (
+                  <TraceEvent
+                    events={[BrowserEvent.onClick]}
+                    name={SharedEventName.NAVBAR_CLICKED}
+                    element={loggingElementName}
+                    key={index}
+                  >
+                    <NavItem onClick={handleNavItemClick} active={currentTab === index} key={key}>
+                      <span>{title}</span>
+                    </NavItem>
+                  </TraceEvent>
+                )
+              })}
+            </Nav>
+          )}
+          {isExplore ? (
+            <FiltersContainer isExplore>
+              <DropdownFilterContainer isExplore>
+                <NetworkFilter />
+                {currentKey === 'tokens' && <TimeSelector />}
+              </DropdownFilterContainer>
+              <SearchContainer isExplore>
+                {currentKey !== 'transactions' && <SearchBar tab={currentKey} />}
+              </SearchContainer>
+            </FiltersContainer>
+          ) : (
+            <>
+              <FiltersContainer isExplore={false}>
+                <NetworkFilter />
+                <TimeSelector />
+              </FiltersContainer>
+              <SearchContainer isExplore={false}>
+                <SearchBar />
+              </SearchContainer>
+            </>
+          )}
+        </NavWrapper>
+        {isExplore ? <Page /> : <TokenTable />}
       </ExploreContainer>
     </Trace>
   )
